==> notebooks/1. Primer dataset-v2.ipynb <==
### 1. dev_notebook_biciMAD stations

This is your development environment...you may start:
- Exploring your data.
- Testing the ad-hoc methods.
- Designing your pipeline.

import pandas as pd
bicimad_df = pd.read_csv('../data/raw/bicimad_stations.csv')
bicimad_df.head()

Al imprimir el dataframe, observo que están delimitados por '\t'. Según la documentación [pandas.read_csv](https://pandas.pydata.org/docs/reference/api/pandas.read_csv.html) debo separar las columnas con `sep='\t':`

bicimad_df = pd.read_csv('../data/raw/bicimad_stations.csv', sep='\t')
bicimad_df.tail(2)

bicimad_df.info()

1. Tenemos **14 columnas.** Solo necesitamos **3** de ellas:

    - **name:** Nombre de la estación de biciMAD
    - **address:** Dirección de la estación de biciMAD
    - **geometry.coordinates:** Coordenadas (longitud y latitud) de la ubicación de biciMAD


2. Tenemos **264 registros** de los cuales debemos revisar en las 3 columnas:

    - Si hay **valores nulos** en las **3 columnas**
    - Si hay **valores únicos** para la **estaciones biciMAD** y coincide con 264
    - Si hay un **máximo** de 264 registros no nulos en la columna `address`
    - Si hay un **máximo** de 264 registros no nulos en la columna `geometry.coordinates`

[Documentación Pandas Dataframes](https://pandas.pydata.org/docs/reference/api/pandas.DataFrame.html)

Contamos los registros de la columna 'name', revisando que haya una estación de biciMAD diferente para cada registro:

**1. Nombres de la estación de biciMAD ['name']**

[pandas.Series.value_counts](https://pandas.pydata.org/docs/reference/api/pandas.Series.value_counts.html): **.value_counts()** regresa una serie con el conteo de valores únicos, en orden descendiente (de mayor a menor). Excluye NA por defecto.

bicimad_df['name'].value_counts()

Para asegurarme, contaré con **.notnull()** los valores **NO nulos**, pasandoles de nuevo un **.value_counts()**

# https://pandas.pydata.org/docs/reference/api/pandas.DataFrame.notnull.html
bicimad_df['name'].notnull().value_counts()

- Tenemos 264 estaciones distintas (sin repetir) de biciMAD, que era lo que estabamos esperando
- NO hay valores nulos, puesto que hay **264 True**

**2. Direcciones de las estaciones de biciMAD ['address']**

Contamos los registros de la columna 'address', revisando que hayan maximo 264 registros en direcciones. Pueden haber repetidos, puesto que puede haber más de una estación de biciMAD por dirección:

bicimad_df['address'].value_counts()

bicimad_df['address'].notnull().value_counts()

- El número de estaciones es **menor o igual a 264 (257)**
- NO hay valores nulos, puesto que hay **264 True**

**3. Coordenadas (longitud y latitud) de la ubicación de biciMAD ['geometry.coordinates']**

bicimad_df['geometry.coordinates'].head(2)

Reviso el primer elemento para tener idea que tipo de datos voy a analizar en la serie bicimad_df['geometry.coordinates']:

print(bicimad_df['geometry.coordinates'][0])
print(type(bicimad_df['geometry.coordinates'][0]))

bicimad_df['geometry.coordinates'].info()

Quiero asegurarme que cada dtype 'object' sea un string como el que he obtenido del primer elemento. 

Para la columna **bicimad_df['geometry.coordinates']** aplico una función lambda a cada uno de los elementos
que he llamado 'coord', que me devuelva **True** con la función **isinstance(),** si todas las coordenadas son
de tipo **str** y que me dé un total de **264.**

- [pandas.DataFrame.apply](https://pandas.pydata.org/docs/reference/api/pandas.DataFrame.apply.html)
- [Python Lambda](https://www.w3schools.com/python/python_lambda.asp)
- [Python isinstance() Function](https://www.w3schools.com/python/ref_func_isinstance.asp)

bicimad_df['geometry.coordinates'].apply(lambda coord: isinstance(coord, str)).value_counts()

Repetiré la misma verificación de las 2 anteriores columnas, con las coordenadas:

bicimad_df['geometry.coordinates'].value_counts()

- Cada coordenada es **única,** dandome un total de **264**
- El conteo de strings dió **264 True**, lo cual NO hay valores nulos

### Modificando el dataframe con las 3 columnas relevantes

bicimad_df = bicimad_df[['name','address','geometry.coordinates']]

bicimad_df.head()

Voy a guardar en una nueva variable **'coordinates_df',** el dataframe con la limpieza de la columna 'geometry.coordinates' con[.str.split()](https://pandas.pydata.org/pandas-docs/stable/reference/api/pandas.Series.str.split.html).

El proceso será el siguiente:

1. Le quito los brackets con **.str.strip('[]')**
2. Elimino las comas con **.str.split(',', expand=True)** para expandir a dos columnas
3. Aplico el método **[.astype('float64')](https://pandas.pydata.org/docs/reference/api/pandas.DataFrame.astype.html)** para que los valores tengan formato float para Pandas

Lo podemos ver a continuación con la nueva variable **'coordinates_df':**

coordinates_df = bicimad_df['geometry.coordinates'].str.strip('[]').str.split(',', expand=True).astype('float64')
coordinates_df

Con **.str.split(',', expand=True)** me separa la columna 'geometry.coordinates' en 2.

0 y 1 para **longitud y latitud,** respectivamente.

El siguiente paso es [renombrar las columnas:](https://pandas.pydata.org/docs/reference/api/pandas.DataFrame.columns.html)

coordinates_df.columns = ['mad_long', 'mad_lat']
coordinates_df

**Concatenar los dataframes por columnas:** agregaré con `axis=1` las columnas de `coordinates_df` al principal `bicimad_df`con [pandas.concat](https://pandas.pydata.org/docs/reference/api/pandas.concat.html)

bicimad_df = pd.concat([bicimad_df,coordinates_df],axis=1)
bicimad_df

Elimino la columna **geometry.coordinates** con [pandas.DataFrame.drop](https://pandas.pydata.org/docs/reference/api/pandas.DataFrame.drop.html)

bicimad_df = bicimad_df.drop(columns=['geometry.coordinates'])

### Dataframe final de bicimad_df

# Renombramos las columnas
bicimad_df = bicimad_df.rename(columns={"name": "BiciMAD station",
                          "address" : "BiciMAD address"})
bicimad_df

bicimad_df.info()

### Exportamos el .csv final de bicimad_stations (v2)

from pathlib import Path
filepath = Path('../data/processed/bicimad_stations.csv')  
filepath.parent.mkdir(parents=True, exist_ok=True) 
bicimad_df.to_csv(filepath)
