==> notebooks/2. Segundo dataset_v2.ipynb <==
### 2. dev_notebook_biciPARKS stations

This is your development environment...you may start:
- Exploring your data.
- Testing the ad-hoc methods.
- Designing your pipeline.

El análisis de este notebook será más resumido, debido a las observaciones rigurosas del primer notebook **(1. dev_notebook_biciMAD stations)** y que son replicados a continuación:

import pandas as pd

biciparks_df = pd.read_csv('../data/raw/bicipark_stations.csv', sep=';') # el delimitador es ';' si abrimos el csv
biciparks_df

De un vistazo, lo que me llamó la atención ha sido la estación **Bicipark Canalejas,** porque no veo información del código postal, **ni** su dirección.

A través de las coordenadas **[40.417807676, -3.6994506]** en Google Maps, encuentro "Parking Centro Canalejas, Cra de S. Jerónimo, 7, 28014 Madrid" **(para coches).** NO es un Bicipark.

He consultado la página de [EMT Madrid](https://www.emtmadrid.es/Aparcamientos/Bicipark/Disponibles?lang=es-ES) y NO existe el **Bicipark Canalejas.** Por lo tanto, me puedo cargar el registro:

biciparks_df = biciparks_df.drop([10])
biciparks_df.tail()

biciparks_df.info()

1. Tenemos **12 columnas.** Solo necesitamos **4** de ellas:

    - **stationName:** Nombre de la estación de biciPARK
    - **address:** Dirección de la estación de biciPARK
    - **zip_code:** Código postal
    - **geometry.coordinates:** Coordenadas (longitud y latitud) de la ubicación de biciPARK

Igual que en el Dataframe BiciMAD, voy a revisar que tipo de dato son algunos elementos de **geometry.coordinates**:

print(biciparks_df['geometry.coordinates'][0])
print(type(biciparks_df['geometry.coordinates'][0]))
print(type(biciparks_df['geometry.coordinates'][5]))

### Modificando el dataframe con las 4 columnas relevantes

biciparks_df = biciparks_df[['stationName','address','zip_code','geometry.coordinates']]

biciparks_df

Haré el mismo tratamiento de biciMAD en la columna **geometry.coordinates:**

coordinates_df = biciparks_df['geometry.coordinates'].str.strip('[]').str.split(',', expand=True).astype('float64')
coordinates_df

coordinates_df.columns = ['park_long', 'park_lat']
coordinates_df

**Concatenar los dataframes por columnas:**

biciparks_df = pd.concat([biciparks_df,coordinates_df],axis=1)
biciparks_df

# Elimino la columna geometry.coordinates
biciparks_df = biciparks_df.drop(columns=['geometry.coordinates'])

### Dataframe final de biciparks_df

# Renombramos las columnas
biciparks_df = biciparks_df.rename(columns={"stationName": "BiciPARK station",
                          "address" : "BiciPARK address"})
biciparks_df

### Exportamos el .csv final de biciparks_df_v2

from pathlib import Path
filepath = Path('../data/processed/biciparks_df.csv')  
filepath.parent.mkdir(parents=True, exist_ok=True) 
biciparks_df.to_csv(filepath)
