==> notebooks/4. Funciones.ipynb <==
### Main Challenge:

You must create a Python App (Data Pipeline) that allow their potential users to find the nearest BiciMAD station to a set of places of interest using the methods included in the module geo_calculations.py.

### Instalaciones accesibles no municipales

Recopila información de todas aquellas instalaciones accesibles no municipales de la ciudad de las que posee conocimiento expreso el Ayuntamiento de Madrid y que están publicadas en el portal institucional madrid.es. 

Incluye, entre otras, información sobre:

-          Ministerios

-          Oficinas de correos

-          Centros de enseñanza y universidades

-          Centros de día y residenciales

-          Museos

-          Espacios de ocio

 

En la información de cada espacio, incluye información (si está disponible) sobre:

- el horario
- los equipamientos de que dispone y/o los servicios que presta
- su dirección completa (incluyendo calle,  distrito, barrio, código postal y coordenadas geográficas)
- los medios de transporte más cercanos disponibles (Bus, metro, Bicimad y Renfe)
- teléfono
- el correo electrónico
- la dirección web donde está publicada en madrid.es para su consulta en un formato más amigable.

**[Enlace](https://datos.madrid.es/portal/site/egob/menuitem.c05c1f754a33a9fbe4b2e4b284f1a5a0/?vgnextoid=f31050abd15b3410VgnVCM1000000b205a0aRCRD&vgnextchannel=374512b9ace9f310VgnVCM100000171f5a0aRCRD&vgnextfmt=default)**

Los métodos incluidos en el módulo geo_calculations.py **mencionados anteriormente** son copiados-pegados textualmente a continuación:

# conda install -c conda-forge geopandas 
# (shapely is a geopandas dependance)

from shapely.geometry import Point # https://geopandas.org/en/stable/docs/reference/api/geopandas.GeoSeries.html
import geopandas as gpd # https://geopandas.org/en/stable/getting_started.html

# Methods

def to_mercator(lat, long):
    # transform latitude/longitude data in degrees to pseudo-mercator coordinates in metres
    c = gpd.GeoSeries([Point(lat, long)], crs=4326)
    c = c.to_crs(3857)
    return c

def distance_meters(lat_start, long_start, lat_finish, long_finish):
    # return the distance in metres between to latitude/longitude pair points in degrees 
    # (e.g.: Start Point -> 40.4400607 / -3.6425358 End Point -> 40.4234825 / -3.6292625)
    start = to_mercator(lat_start, long_start)
    finish = to_mercator(lat_finish, long_finish)
    return start.distance(finish)

Sin embargo, tuve problemas para instalar Geopandas. Así que tuve que hacerlo desde el Notebook con las siguientes líneas de comandos:

- `%pip install geopandas`
- `%pip install shapely`

Esta es la [documentación](https://geopandas.org/en/stable/docs/reference/geoseries.html) de **Geoseries** en Pandas.

Por supuesto, importamos las siguientes **librerías:**

import pandas as pd
import numpy as np

#### Lo que generamos de los notebooks anteriores, importamos las versiones de los siguientes .cvs:

- bicimad_df
- biciparks_df
- inst_no_munic_df

#### BiciMAD:

bicimad_df = pd.read_csv('../data/processed/bicimad_stations.csv')
bicimad_df = bicimad_df.rename(columns={"mad_long": "BiciMAD Long.",
                                        "mad_lat" : "BiciMAD Lat."})
bicimad_df = bicimad_df.drop(columns='Unnamed: 0')
bicimad_df

#### Biciparks:

biciparks_df = pd.read_csv('../data/processed/biciparks_df.csv')
biciparks_df = biciparks_df.rename(columns={"park_long": "BiciPARK Long.",
                                            "park_lat" : "BiciPARK Lat."})
biciparks_df = biciparks_df.drop(columns=['Unnamed: 0','zip_code'])
biciparks_df

#### Instalaciones Accesibles NO municipales:

inst_no_munic_df = pd.read_csv('../data/processed/inst_no_munic_df.csv')
inst_no_munic_df = inst_no_munic_df.drop(columns='Unnamed: 0')
inst_no_munic_df = inst_no_munic_df.drop(columns='Postal code')
inst_no_munic_df = inst_no_munic_df.rename(columns={"place_lat": "Place LAT.",
                                            "place_long" : "Place LONG."})
inst_no_munic_df

El output debe lucir algo ["así"](https://github.com/ih-datapt-mad/ih_datamadpt0923_project_m1)

Vamos a usar la función para **calcular distancias** entre coordenadas:

def distance_meters(lat_start, long_start, lat_finish, long_finish):
    # return the distance in metres between to latitude/longitude pair points in degrees 
    # (e.g.: Start Point -> 40.4400607 / -3.6425358 End Point -> 40.4234825 / -3.6292625)
    start = to_mercator(lat_start, long_start)
    finish = to_mercator(lat_finish, long_finish)
    return start.distance(finish)

Haremos una primera prueba desde **BiciMAD** 1a - Puerta del Sol A, calculando su distancia hasta **Sede QuedaT.com Chamartín**

bicimad_df.head(1)

inst_no_munic_df.head(1)

# distance_meters(lat_start, long_start, lat_finish, long_finish)

# El punto inicial es Puerta del Sol (lat_start, long_start) y el destino es Chamartín (lat_finish, long_finish)

distance_meters(40.417214,-3.701834, 40.447907,-3.663171)

La primera prueba nos dice que hay una distancia de 5502 mts (5.5 kms) **desde** Bicimad de Puerta del Sol **hasta** Sede QuedaT.com Chamartín.

### Combinando Dataframes con [Pandas Cross-join](https://linuxhint.com/pandas-cross-join/)

#### 1. Entre BiciMAD y los Sitios de interés

# Creamos una columna 'key' en bicimad_df y en inst_no_munic_df para tener una llave en común 
bicimad_df['key'] = 0
inst_no_munic_df['key'] = 0

# Creamos un nuevo df bicimad_project_df, haciendo el merge de bicimad con el de los sitios de interés
bicimad_project_df = bicimad_df.merge(inst_no_munic_df, on='key', how='outer')
bicimad_project_df.head()

# Combinando cada fila de las estaciones bicimad con cada una de los sitios de interés, obtenemos 107712 filas
len(bicimad_project_df)

#### 2. Entre Biciparks y los Sitios de interés

# Creamos una columna 'key' en biciparks_df y en inst_no_munic_df para tener una llave en común 
biciparks_df['key'] = 0

# Creamos un nuevo df parks_project_df, haciendo el merge de biciparks con el de los sitios de interés
parks_project_df = biciparks_df.merge(inst_no_munic_df, on='key', how='outer')
parks_project_df.head()

# Combinando cada fila de las estaciones biciparks con cada una de los sitios de interés, obtenemos 4080 filas
len(parks_project_df)

### Calculando distancias entre coordenadas con def distance_meters

#### 1. Distancia en metros desde BiciMAD de destino hasta cada Sitio de interés

# distance_meters(lat_start, long_start, lat_finish, long_finish)
bicimad_project_df['Distance from station in mts'] = round(bicimad_project_df.apply(lambda row: distance_meters(row['BiciMAD Lat.'], row['BiciMAD Long.'], row['Place LAT.'], row['Place LONG.']), axis=1), 2)

# Ordenando de menor a mayor distancia
bicimad_project_df = bicimad_project_df.sort_values(by='Distance from station in mts')

(La función se tomó **10 minutos** en calcular las distancias de todas las filas)

# Cambiando el orden de las columnas
column_order = ['Place of interest','Place address','BiciMAD station', 'BiciMAD address','Distance from station in mts']
bicimad_project_df = bicimad_project_df[column_order]
bicimad_project_df = bicimad_project_df.sort_values(by='Distance from station in mts')
bicimad_project_df

bicimad_project_df.info()

#### 2. Distancia en metros desde Biciparks de destino hasta cada Sitio de interés

parks_project_df['Distance from parking in mts'] = round(parks_project_df.apply(lambda row: distance_meters(row['BiciPARK Lat.'], row['BiciPARK Long.'], row['Place LAT.'], row['Place LONG.']), axis=1), 2)
parks_project_df.sort_values(by='Distance from parking in mts')

# Cambiando el orden de las columnas
column_order = ['Place of interest','Place address','BiciPARK station', 'BiciPARK address','Distance from parking in mts']
parks_project_df = parks_project_df[column_order]
parks_project_df = parks_project_df.sort_values(by='Distance from parking in mts')
parks_project_df

parks_project_df.info()

# Distancias cortas a través de un groupby()

#### 1. BiciMAD

min_bicimad_df = bicimad_project_df.groupby('Place of interest', as_index=False).agg({'Place address': 'first', 'Distance from station in mts': 'min'})
min_bicimad_df

# Pruebas
min_bicimad_df[min_bicimad_df['Place of interest'] == 'Colegio Oficial de Arquitectos de Madrid (COAM)']

# Creamos un nuevo df bicimad_project_df, haciendo el merge de bicimad con el de los sitios de interés
bicimad_project_df = min_bicimad_df.merge(bicimad_project_df, on=['Place of interest', 'Place address', 'Distance from station in mts'], how='inner')

bicimad_project_df

# Incluir columna Type of place (*) 'Instalaciones accesibles NO municipales'

bicimad_project_df.insert(loc=1, column='Type of place', value='Instalaciones accesibles NO municipales')

# df.insert(loc=1, column='NuevaColumna', value=[1, 2, 3, 4]) => https://docs.kanaries.net/es/topics/Pandas/pandas-add-column

# Bicimad Dataframe Final

bicimad_project_df

### Exportando .csv BiciMAD_Final

from pathlib import Path
filepath = Path('../data/results/bicimad_project_df.csv', index=False)  
filepath.parent.mkdir(parents=True, exist_ok=True) 
bicimad_project_df.to_csv(filepath, index=False, encoding='utf-8')

#### 2. BiciPARKS

min_biciparks_df = parks_project_df.groupby('Place of interest', as_index=False).agg({'Place address': 'first', 'Distance from parking in mts': 'min'})
min_biciparks_df

# Creamos un nuevo df parks_project_df, haciendo el merge de bicimad con el de los sitios de interés
parks_project_dfparks_project_df = min_biciparks_df.merge(parks_project_df, on=['Place of interest', 'Place address', 'Distance from parking in mts'], how='inner')

parks_project_df

parks_project_df.insert(loc=1, column='Type of place', value='Instalaciones accesibles NO municipales')

# Biciparks Dataframe Final

parks_project_df

### Exportando .csv Bicipark_Final

from pathlib import Path
filepath = Path('../data/results/biciparks_project_df.csv', index=False)  
filepath.parent.mkdir(parents=True, exist_ok=True) 
parks_project_df.to_csv(filepath, index=False, encoding='utf-8')

## Objetivo: Teniendo los 2 dataframes:

It must provide, **at least,** two options for the final user to select when executing using argparse:
- (1) To get the table for every 'Place of interest' included in the dataset (or a set of them)
- (2) To get the table for a specific 'Place of interest' imputed by the user.

**Fuentes**
- [Usar Argparse en Python](https://www.youtube.com/watch?v=ng-RYmDyE5A&ab_channel=episuarez)

# Reciclaje de códigos

### Exportamos el .csv final de bicimad_project_df y parks_project_df

# from pathlib import Path
# filepath = Path('./modules/bicimad_project_df.csv')  
# filepath.parent.mkdir(parents=True, exist_ok=True) 
# bicimad_project_df.to_csv(filepath)

# from pathlib import Path
# filepath = Path('./modules/parks_project_df.csv')  
# filepath.parent.mkdir(parents=True, exist_ok=True) 
# parks_project_df.to_csv(filepath)

### Argparse

[argparse — Analizador sintáctico (Parser) para las opciones, argumentos y sub-comandos de la línea de comandos](https://docs.python.org/es/3/library/argparse.html)

# from argparse import ArgumentParser

# parser = ArgumentParser(description="'get the table for every 'Place of interest' included in.')
# parser.add_argument('integers', metavar='N', type=int, nargs='+',
#                     help='an integer for the accumulator')
# parser.add_argument('--sum', dest='accumulate', action='store_const',
#                     const=sum, default=max,
#                     help='sum the integers (default: find the max)')

# args = parser.parse_args()
# print(args.accumulate(args.integers))

# import argparse

# parser = argparse.ArgumentParser(description='Process some integers.')
# parser.add_argument('integers', metavar='N', type=int, nargs='+',
#                     help='an integer for the accumulator')
# parser.add_argument('--sum', dest='accumulate', action='store_const',
#                     const=sum, default=max,
#                     help='sum the integers (default: find the max)')

# args = parser.parse_args()
# print(args.accumulate(args.integers))

`module geo_calculations.py.`
[Geopandas library](https://geopandas.org/)

# bicimad_df = pd.read_csv('./data/bicimad_stations.csv')
# bicimad_df.head()

# biciparks_df = pd.read_csv('./data/biciparks_df_v2.csv')
# inst_no_munic_d
# f = pd.read_csv('./data/inst_no_munic_df_v2.csv')

# bicimad_df = bicimad_df.drop(columns='Unnamed: 0')
# bicimad_df = bicimad_df.rename(columns={"mad_long": "BiciMAD Long.",
#                                         "mad_lat" : "BiciMAD Lat."})
# bicimad_df

# biciparks_df = biciparks_df.drop(columns='Unnamed: 0')
# biciparks_df = biciparks_df.drop(columns='zip_code')
# biciparks_df = biciparks_df.rename(columns={"park_long": "BiciPARK Long.",
#                                             "park_lat" : "BiciPARK Lat."})
# biciparks_df

# inst_no_munic_df = inst_no_munic_df.drop(columns='Unnamed: 0')
# inst_no_munic_df = inst_no_munic_df.drop(columns='Postal code')
# inst_no_munic_df = inst_no_munic_df.rename(columns={"place_lat": "Place LAT.",
#                                             "place_long" : "Place LONG."})
# # inst_no_munic_df

# https://www.geeksforgeeks.org/python-program-to-perform-cross-join-in-pandas/

# project_df['distance_meters'] = round(project_df.apply(lambda row: distance_meters(row['BiciMAD Lat.'], row['BiciMAD Long.'], row['Place LAT.'], row['Place LONG.']), axis=1), 2)
# project_df.sort_values(by='distance_meters')

# bicimad_project_df.groupby('Place of interest', as_index=False).agg('Place address',{'distance_meters': ['mean', 'min', 'max']})
# bicimad_project_df.groupby('Place of interest', as_index=False).agg({'Place address': 'first', 'distance_meters': ['mean', 'min', 'max']})

# inst_no_munic_df['Place address'] = inst_no_munic_df['Place address'].apply(lambda row: row.replace('&amp;Ntilde;','Ñ'))

# sitios_interes = bicistation_df['Place of interest'].apply(lambda row: row.replace('&amp;Ntilde;','Ñ')).tolist()

# sitios_interes = [row for row in bicistation_df['Place of interest'] if args.lugar in row].tolist()
# sitios_interes = [row for row in bicistation_df['Place of interest'] if args.lugar in row].tolist()

**Octavio:**

Crear una app de usuario en silla de ruedas que necesita ir a un sitio.

Meterle el lugar de destino y ver si está adaptado a acceso de movilidad reducida.
