==> notebooks/5. Bicis por estación.ipynb <==
import pandas as pd

# 1. BiciMAD: Bicis por estación

bicimad_df = pd.read_csv('../data/raw/bicimad_stations.csv',sep='\t')
bicimad_df.head()

bicimad_df['no_available'].value_counts()

**Columnas relevantes:**

- **BiciMAD station (name):** 1a - Puerta del Sol A
- **Bikes available	(no_available):**  1	

bicimad_df = bicimad_df.rename(columns={"name": "BiciMAD station",
                                        "no_available" : "Bikes available"})

# bicimad_df['name'].value_counts()

columns = ['BiciMAD station','Bikes available']
bicimad_dispo = bicimad_df[columns]

bicimad_dispo

bicimad_df = pd.read_csv('../data/results/bicimad_project_df.csv')
bicimad_df.head(2)

bicimad_project_df_test = bicimad_dispo.merge(bicimad_df, on=['BiciMAD station'], how='right')

column_order = ['Place of interest','Type of place','Place address','Distance from station in mts','BiciMAD station','BiciMAD address','Bikes available']
bicimad_df = bicimad_project_df_test[column_order]

bicimad_df.head(2)

bicimad_df.info()

# Exportando csv

from pathlib import Path
filepath = Path('../data/results/bicimad_vfinal_df.csv', index=False)  
filepath.parent.mkdir(parents=True, exist_ok=True) 
bicimad_df.to_csv(filepath, index=False, encoding='utf-8')

# 2. Biciparks: Puestos disponibles por estación

biciparks_df = pd.read_csv('../data/raw/bicipark_stations.csv', sep=';') # el delimitador es ';' si abrimos el csv
biciparks_df.head(2)

**Columnas relevantes:**

- **stationName (BiciPARK station):** Bicipark Fuente de la Mora
- **free_places (Spots available):** 9

biciparks_df = biciparks_df.rename(columns={"stationName": "BiciPARK station",
                                        "free_places" : "Spots available"})

# bicimad_df['name'].value_counts()

columns = ['BiciPARK station','Spots available']
spots_bicipark = biciparks_df[columns]
spots_bicipark = spots_bicipark.drop(10)

spots_bicipark

spots_bicipark['Spots available'] = spots_bicipark['Spots available'].abs()

spots_bicipark

biciparks_df = pd.read_csv('../data/results/biciparks_project_df.csv')
biciparks_df.head(2)

bicipark_project_df_test = spots_bicipark.merge(biciparks_df, on=['BiciPARK station'], how='right')

bicipark_project_df_test.head(1)

column_order = ['Place of interest','Type of place','Place address','Distance from parking in mts','BiciPARK station','BiciPARK address','Spots available']
bicipark_df = bicipark_project_df_test[column_order]

bicipark_df.head()

# Exportando csv

from pathlib import Path
filepath = Path('../data/results/biciparks_vfinal_df.csv', index=False)  
filepath.parent.mkdir(parents=True, exist_ok=True) 
bicipark_df.to_csv(filepath, index=False, encoding='utf-8')

# Exportando csv

from pathlib import Path
filepath = Path('../modules/biciparks_project_df_v4.csv', index=False)  
filepath.parent.mkdir(parents=True, exist_ok=True) 
bicipark_df.to_csv(filepath, index=False, encoding='utf-8')
